pyproject.toml
[tool.poetry]
name = "pytest-fastapi-crud-example"
version = "0.1.0"
description = "Example of testing a FastAPI CRUD API in Pytest"
authors = ["ericsalesdeandrade <sdaeric19@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.110.3"
uvicorn = "^0.29.0"
pydantic = "^2.7.1"
httpx = "^0.27.0"
sqlalchemy-utils = "^0.41.2"
psycopg2-binary = "^2.9.9"
fastapi-cli = "^0.0.7"
python-jose = "^3.4.0"
google-auth = "^2.38.0"
google-auth-httplib2 = "^0.2.0"
requests = "^2.32.3"


[tool.poetry.group.test.dependencies]
pytest = "^8.2.0"
pytest-randomly = "^3.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
